sistema-edah/src/pyproject.toml
[project]
name = "django-sis-igreja"
version = "0.1.0"
description = "Sistema Django para gestão de igreja"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.0,<6",
    "djangorestframework>=3.15,<4",
    "djangorestframework-simplejwt>=5.3,<6",
    "django-cors-headers>=4.3,<5",
    "firebase-admin>=6.5,<7",
    # Multi-tenancy SaaS (requerido quando DATABASE_URL estiver configurada)
    "django-tenants>=3.6,<4",
    "psycopg2-binary>=2.9,<3",
    "django-environ>=0.11,<1",
    "whitenoise>=6.6,<7",
]

[dependency-groups]
dev = [
    "pip-audit>=2.9,<3",
    "pytest>=8.0,<9",
    "pytest-cov>=5.0,<6",
    "pytest-django>=4.8,<5",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py", "tests.py", "tests_*.py"]
testpaths = ["tests", "church"]
addopts = "-q --strict-markers --maxfail=1 --cov=church --cov-report=term-missing --cov-fail-under=70"
markers = [
    "critico: testes de fluxos criticos de negocio",
    "integracao: testes de integracao entre camadas e fronteiras internas",
]
